targets: Add all copies to the FileWrites item list
The FileWrites item list is used by MSBuild to keep track of all files
that have been written during the build, and need to be cleaned up in
the Clean target. Adding the copied files to them ensures that they are
removed when the Clean target is run.

For more infos see http://msdn.microsoft.com/en-us/magazine/dd419659.aspx#id0090093